+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Epaminondas
   - Fabrik
   - Fendo
+  - Lines of Action
 
 ## [0.4.0] - 2021-11-15